# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GraphDB API",
    "description": "GraphDB network service layer (HTTP/gRPC)",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/statistics/sessions/{id}": {
      "get": {
        "tags": [
          "Statistics"
        ],
        "summary": "Obtaining session statistics",
        "operationId": "session",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Session id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session statistics",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
...
    "/v1/users": {
      "get": {
        "tags": [
          "Auth"
        ],
        "operationId": "get_v1_users",
        "responses": {
          "200": {
            "description": "User list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal error"
          }
        }
      },
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "post_v1_users",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "User created"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
    "/v1/users/{name}": {
      "delete": {
        "tags": [
          "Auth"
        ],
        "operationId": "delete_v1_users_name",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Username",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
    "/v1/users/{name}/disable": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "post_v1_users_name_disable",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Username",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User disabled"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
    "/v1/users/{name}/enable": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "post_v1_users_name_enable",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Username",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User enabled"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
    "/v1/users/{name}/grant": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "post_v1_users_name_grant",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Username",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Role granted"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
    "/v1/users/{name}/password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "post_v1_users_name_password",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Username",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Password reset"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
    "/v1/users/{name}/revoke": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "post_v1_users_name_revoke",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Username",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Role revoked"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateUserRequest": {
        "type": "object",
        "description": "Create user request for the admin channel.",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
...
      "GrantRoleRequest": {
        "type": "object",
        "description": "Grant role request for the admin channel.",
        "required": [
          "role",
          "space"
        ],
        "properties": {
          "role": {
            "type": "string"
          },
          "space": {
            "type": "string"
          }
        }
      },
...
      "ResetPasswordRequest": {
        "type": "object",
        "description": "Reset password request for the admin channel.",
        "required": [
          "password"
        ],
        "properties": {
          "password": {
            "type": "string"
          }
        }
      },
...
      "RevokeRoleRequest": {
        "type": "object",
        "description": "Revoke role request for the admin channel.",
        "required": [
          "space"
        ],
        "properties": {
          "space": {
            "type": "string"
          }
        }
      },
...
      "UserListItem": {
        "type": "object",
        "description": "User list entry for the management view.",
        "required": [
          "username"
        ],
        "properties": {
          "last_active": {
            "type": [
              "string",
              "null"
            ]
          },
          "role": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": [
              "string",
              "null"
            ]
          },
          "username": {
            "type": "string"
          }
        }
      },
      "UserListResponse": {
        "type": "object",
        "description": "User list response for the management view.",
        "properties": {
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserListItem"
            }
          }
        }
      },
...
    }
  },
...
}